Rename molecule_moderator folder to moleculeModerator
