+++ app/packs/entrypoints/application.js
@@ -7,7 +7,7 @@ var AdminHome = require('src/apps/admin');
 var ChemScanner = require('src/apps/chemscanner/');
 var ChemSpectra = require('src/apps/chemspectra/ChemSpectra');
 var ChemSpectraEditor = require('src/apps/chemspectra/ChemSpectraEditor');
-var MoleculeModerator = require('src/apps/molecule_moderator');
+var MoleculeModerator = require('src/apps/moleculeModerator');
 var OmniauthCredential = require('src/apps/omniauthCredential');
 var UserCounter = require('src/apps/user_counter');
 var ScifinderCredential = require('src/apps/scifinder_credential');
