+++ app/packs/src/apps/moleculeModerator/MoleculeModeratorComponent.js
